Fixes the filter icon support
Since the java Drid doesn't have support for custom column's header renderer
anymore, I used a somewhat radical approach to the problem: I monkey patched
the function that sets the client side header renderer. If a column has a
filter, then we will use enhanced-grid-sorter in the header. If the column
doesn't have, then the default implementation will be used.